+++ src/nexus/dock/vina/pipeline.py
@@ -1,7 +1,6 @@
 from dataclasses import dataclass
 
 from nexus.dock.dock_config import DockConfig
-from nexus.prep.lig.ligands_prep import ligands_prep
 from nexus.dock.vina._prep_rec import vina_prep_rec
 from nexus.dock.utils.matchmixer import matchmixer
 from nexus.dock.vina._docking import vina_docking
@@ -14,15 +13,15 @@ class VinaPipeline():
 
     def run(self):
         self.dcfg.common.program = "vina"
+        if ".pdbqt" not in self.dcfg.ligands.suffix:
+            raise ValueError("Ligands for Vina must have '.pdbqt' suffix.")
+        lig_paths = self.dcfg.ligands.source
+        rec_bundles = vina_prep_rec(self.dcfg)
 
-        prepped_ligs = ligands_prep(self.dcfg)
-        prepped_recs = vina_prep_rec(self.dcfg)
-
-        pairs = matchmixer(prepped_recs, prepped_ligs, 
-                           self.dcfg.common.prepared_suffix, self.dcfg.common.mode)
-
+        pairs = matchmixer(rec_bundles, lig_paths, 
+                           "self.dcfg.common.prepared_suffix", self.dcfg.common.mode)
+        # Disable validate for now
         out_files = vina_docking(self.dcfg, pairs)
-        docking_summary = write_summary_csv(self.dcfg, out_files, prepped_recs)
+        docking_summary = write_summary_csv(self.dcfg, out_files, rec_bundles)
 
-        config_files = [bundle.vina_config for bundle in prepped_recs]
-        final_copy(self.dcfg, prepped_recs, docking_summary, out_files, config_files)
+        final_copy(self.dcfg, rec_bundles, docking_summary, out_files)