+++ src/nexus/prep/mutate/pipeline.py
@@ -1,5 +1,6 @@
+from pathlib import Path
 from pydantic import BaseModel
-from nexus.prep.prep_config import PrepConfig, default_output
+from nexus.prep.prep_config import PrepConfig
 from nexus.core.extract_files import extract_files
 from nexus.prep.mutate._chimerax_mutate import chimerax_mutate
 
@@ -25,7 +26,8 @@ class MutatePipeline(BaseModel):
         if ".pdb" not in self.pcfg.common.suffix and ".cif" not in self.pcfg.common.suffix:
             raise ValueError("Output receptor format must be 'pdb' or 'cif'.")
 
-        self.pcfg = default_output(self.pcfg)
+        if self.pcfg.common.output_dir is None:
+            self.pcfg.common.output_dir = Path.cwd()
 
         chimerax_mutate(self.pcfg)