+++ src/nexus/prep/prep_config.py
@@ -39,15 +39,3 @@ def load_prep_config(path):
         pcfg.common.output = pcfg.common.input.parent
 
     return pcfg
-
-
-def default_output(pcfg: PrepConfig):
-    if pcfg.common.output_dir is None:
-        if isinstance(pcfg.common.input, Path):
-            pcfg.common.output_dir = pcfg.common.input.parent
-            pcfg.common.input = [pcfg.common.input]
-        elif isinstance(pcfg.common.input, list):   
-            pcfg.common.output_dir = pcfg.common.input[0].parent
-    pcfg.common.output_dir.mkdir(parents=True, exist_ok=True) 
-
-    return pcfg